+++ R/map_layer_scatterplot.R
@@ -161,8 +161,8 @@ add_scatterplot <- function(
 	l[["radius"]] <- force(radius)
 	l[["fill_colour"]] <- force(fill_colour)
 	l[["fill_opacity"]] <- resolve_opacity(fill_opacity)
-	l[["stroke_colour"]] <- if(!is.null( stroke_colour ) ) { force(stroke_colour) } else { force( fill_colour ) }
-	l[["stroke_opacity"]] <- if(!is.null( stroke_opacity ) ) { resolve_opacity( stroke_opacity ) } else { resolve_opacity( fill_opacity ) }
+	l[["stroke_colour"]] <- force( stroke_colour)
+	l[["stroke_opacity"]] <- force( stroke_opacity )
 	l[["stroke_width"]] <- force( stroke_width )
 	l[["tooltip"]] <- force(tooltip)
 	l[["id"]] <- force(id)