+++ R/mapdeck_dependencies.R
@@ -169,11 +169,13 @@ deckgl_min_js <- function() {
 	)
 }
 
+## https://api.mapbox.com/mapbox-gl-js/v1.9.0/mapbox-gl.js
+## https://api.mapbox.com/mapbox-gl-js/v1.9.0/mapbox-gl.css
 mapboxgl <- function() {
 	list(
 		createHtmlDependency(
 			name = "mapboxgl",
-			version = "0.52.0",
+			version = "1.9.0",
 			src = system.file("htmlwidgets/lib/", package = "mapdeck"),
 			script = c("mapbox-gl.js"),
 			stylesheet = c("mapbox-gl.css"),