+++ R/mapdeck_map.R
@@ -28,6 +28,9 @@
 #'   \item{viewBounds}
 #'   \item{interactionState}
 #' }
+#'
+#' @param repeat_view Logical indicating if the layers should repeat at low zoom levels
+#'
 #' @section Access Tokens:
 #'
 #' If the \code{token} argument is not used, the map will search for the token, firstly by
@@ -50,7 +53,8 @@ mapdeck <- function(
 	zoom = 0,
 	bearing = 0,
 	location = c(0, 0),
-	show_view_state = FALSE
+	show_view_state = FALSE,
+	repeat_view = FALSE
 	) {
 
   # forward options using x
@@ -62,6 +66,7 @@ mapdeck <- function(
     , location = force( as.numeric( location ) )
     , bearing = force( bearing )
     , show_view_state = force( show_view_state )
+    , repeat_view = force( repeat_view )
   )
 
   # deps <- list(