+++ tests/testthat/test-layer_polygon.R
@@ -3,7 +3,7 @@ context("polygon")
 
 test_that("add_polygon accepts multiple objects", {
 
-	# testthat::skip_on_cran()
+	testthat::skip_on_cran()
 	# testthat::skip_on_travis()
 
 	# library(sf)